requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
cached-property==1.5.2
certifi==2024.2.2
charset-normalizer==2.0.12
click==8.0.3
exceptiongroup==1.2.0
expiringdict==1.2.1
fastapi==0.111.0
frozenlist==1.4.0
h11==0.110.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
nextcord==2.6.0
openpyxl==3.1.2
pydantic==1.6
pydantic-core==0.0.21
python-baseconv==1.2.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.26.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.8.0
urllib3==1.26.7
uvicorn==0.29.0
Werkzeug==3.0.1
yarl==1.9.3